move composition functions to composition.py
