+++ vivarium/compartment/process.py
@@ -469,209 +469,3 @@ class Compartment(Store):
             'data': data}
 
         self.emitter.emit(emit_config)
-
-
-
-
-## functions for testing
-def toy_composite(config):
-    '''
-    a toy composite function for testing
-    returns a dictionary with 'processes', 'states', and 'options'
-
-    '''
-
-    # toy processes
-    class ToyMetabolism(Process):
-        def __init__(self, initial_parameters={}):
-            ports = {'pool': ['GLC', 'MASS']}
-            parameters = {'mass_conversion_rate': 1}
-            parameters.update(initial_parameters)
-
-            super(ToyMetabolism, self).__init__(ports, parameters)
-
-        def next_update(self, timestep, states):
-            update = {}
-            glucose_required = timestep / self.parameters['mass_conversion_rate']
-            if states['pool']['GLC'] >= glucose_required:
-                update = {
-                    'pool': {
-                        'GLC': -2,
-                        'MASS': 1}}
-
-            return update
-
-    class ToyTransport(Process):
-        def __init__(self, initial_parameters={}):
-            ports = {
-                'external': ['GLC'],
-                'internal': ['GLC']}
-            parameters = {'intake_rate': 2}
-            parameters.update(initial_parameters)
-
-            super(ToyTransport, self).__init__(ports, parameters)
-
-        def next_update(self, timestep, states):
-            update = {}
-            intake = timestep * self.parameters['intake_rate']
-            if states['external']['GLC'] >= intake:
-                update = {
-                    'external': {'GLC': -2, 'MASS': 1},
-                    'internal': {'GLC': 2}}
-
-            return update
-
-    class ToyDeriveVolume(Process):
-        def __init__(self, initial_parameters={}):
-            ports = {
-                'compartment': ['MASS', 'DENSITY', 'VOLUME']}
-            parameters = {}
-
-            super(ToyDeriveVolume, self).__init__(ports, parameters)
-
-        def next_update(self, timestep, states):
-            volume = states['compartment']['MASS'] / states['compartment']['DENSITY']
-            update = {
-                'compartment': {'VOLUME': volume}}
-
-            return update
-
-    class ToyDeath(Process):
-        def __init__(self, initial_parameters={}):
-            ports = {
-                'compartment': ['VOLUME'],
-                'global': ['processes']}
-            super(ToyDeath, self).__init__(ports, {})
-
-        def next_update(self, timestep, states):
-            volume = states['compartment']['VOLUME']
-            update = {}
-
-            if volume > 1.0:
-                # kill the cell
-                update = {
-                    'global': {
-                        'processes': []}}
-
-            return update
-
-
-    processes = [
-        {'metabolism': ToyMetabolism(
-            initial_parameters={
-                'mass_conversion_rate': 0.5}), # example of overriding default parameters
-         'transport': ToyTransport()},
-        {'external_volume': ToyDeriveVolume(),
-         'internal_volume': ToyDeriveVolume()},
-        {'death': ToyDeath()}]
-
-    # declare the states
-    states = {
-        'periplasm': Store(
-            initial_state={'GLC': 20, 'MASS': 100, 'DENSITY': 10, 'VOLUME': 100/10},
-            schema={
-                'VOLUME': {
-                    'updater': 'set'}}),
-        'cytoplasm': Store(
-            initial_state={'MASS': 3, 'DENSITY': 10, 'VOLUME': 3/10},
-            schema={
-                'VOLUME': {
-                    'updater': 'set'}})}
-
-    # hook up the ports in each process to compartment states
-    topology = {
-        'metabolism': {
-            'pool': 'cytoplasm'},
-        'transport': {
-            'external': 'periplasm',
-            'internal': 'cytoplasm'},
-        'death': {
-            'compartment': 'cytoplasm',
-            'global': COMPARTMENT_STATE},
-        'external_volume': {
-            'compartment': 'periplasm'},
-        'internal_volume': {
-            'compartment': 'cytoplasm'}}
-
-    # emitter that prints to the terminal
-    emitter = emit.get_emitter({
-        'type': 'print',
-        'keys': {
-            'periplasm': ['GLC', 'MASS'],
-            'cytoplasm': ['MASS']}})
-
-    # schema for states
-    schema = {}
-
-    options = {
-        # 'environment_port': 'environment',
-        # 'exchange_port': 'exchange',
-        'schema': schema,
-        'emitter': emitter,
-        'topology': topology,
-        'initial_time': 0.0}
-
-    return {
-        'processes': processes,
-        'states': states,
-        'options': options}
-
-def test_compartment(composite=toy_composite):
-    compartment = load_compartment(composite)
-    settings = {
-        'timestep': 1,
-        'total_time': 20}
-
-    saved_state = simulate_compartment(compartment, settings)
-    return saved_state
-
-def load_compartment(composite=toy_composite, boot_config={}):
-    '''
-    put a composite function into a compartment
-
-    inputs:
-        - composite is a function that returns a dict with 'processes', 'states', and 'options'
-        for configuring a compartment
-        - boot_config (dict) with specific parameters for the processes
-    return:
-        - a compartment object for testing
-    '''
-
-    composite_config = composite(boot_config)
-    processes = composite_config['processes']
-    states = composite_config['states']
-    options = composite_config['options']
-    options.update({
-        'emitter': boot_config.get('emitter', 'timeseries'),
-        'time_step': boot_config.get('time_step', 1.0)})
-
-    return Compartment(processes, states, options)
-
-def simulate_compartment(compartment, settings={}):
-    '''
-    run a compartment simulation
-    '''
-
-    timestep = settings.get('timestep', 1)
-    total_time = settings.get('total_time', 10)
-
-    # data settings
-    emit_timeseries = settings.get('emit_timeseries', False)
-
-    # run simulation
-    time = 0
-    while time < total_time:
-        time += timestep
-        compartment.update(timestep)
-
-    if emit_timeseries:
-        return compartment.emitter.get_timeseries()
-    else:
-        return compartment.emitter.get_data()
-
-
-if __name__ == '__main__':
-    saved_state = test_compartment()
-    for time, state in saved_state.items():
-        print('{}: {}'.format(time,state))
-
