+++ vivarium/composites/flagella_expression.py
@@ -1,6 +1,6 @@
 import os
 
-from vivarium.compartment.process import load_compartment, simulate_compartment
+from vivarium.compartment.composition import load_compartment, simulate_compartment
 from vivarium.data.nucleotides import nucleotides
 from vivarium.data.amino_acids import amino_acids
 from vivarium.data.chromosomes.flagella_chromosome import FlagellaChromosome
