+++ vivarium/composites/master.py
@@ -3,12 +3,11 @@ from __future__ import absolute_import, division, print_function
 import os
 
 from vivarium.compartment.process import (
-    initialize_state,
-    load_compartment)
+    initialize_state)
 from vivarium.compartment.composition import (
     get_derivers,
     simulate_with_environment,
-    plot_simulation_output)
+    plot_simulation_output, load_compartment)
 from vivarium.composites.gene_expression import plot_gene_expression_output
 
 # processes
